Fold render into processor.ts; move the reporter to scripts/ with typecheck
Per review: renderFlakeDashboardIssue now lives in processor.ts (the
ai-linter.ts layout — processor + its mechanical publisher in one file),
and the CI reporter moves to packages/iterate/scripts/ where knip's
existing glob covers it. The package tsconfig now includes scripts/**
so the reporter AND build.ts are typechecked (tsgo --noEmit) — that dir
previously compiled unchecked.
