+++ packages/iterate/tsconfig.json
@@ -21,6 +21,6 @@
     "rootDir": "../..",
     "skipLibCheck": true
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx"],
+  "include": ["scripts/**/*.ts", "src/**/*.ts", "src/**/*.tsx"],
   "exclude": ["node_modules", "dist"]
 }
